+++ src/dg_scripts.cpp
@@ -1525,7 +1525,7 @@ int text_processed(char *field, char *subfield, struct trig_var_data *vd, char *
 
 	if (!str_cmp(field, "strlen"))  	// strlen
 	{
-		sprintf(str, "%d", strlen(vd->value));
+		sprintf(str, "%lu", static_cast<unsigned long>(strlen(vd->value)));
 		return TRUE;
 	}
 	else if (!str_cmp(field, "trim"))  	// trim
